requirements.txt
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
Flask==2.3.2
google-ai-generativelanguage==0.1.0
google-api-core==2.11.1
google-auth==2.20.0
google-generativeai==0.1.0rc1
googleapis-common-protos==1.59.1
grpcio==1.54.2
grpcio-status==1.54.2
idna==3.4
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
proto-plus==1.22.2
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
python-dotenv==1.0.0
requests==2.31.0
rsa==4.9
six==1.16.0
urllib3==1.26.16
Werkzeug==2.3.6
zipp==3.15.0
